+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.14.0 - (Cool Grey Vienna) - June 2021
+
+### Bugfixes and Improvements 3.14.0
+
+- CLI: new `mc asset-lock` command [#271]
+- SDK: AssetManagement `GetModelLock` and `PutModelLock` Methods
+- Bumped all dependencies
+
 ## 3.13.0 - (Phlox Vienna) - May 2021
 
 ### Bugfixes and Improvements 3.13.0